+++ internal/format/all/all.go
@@ -11,4 +11,5 @@ import (
 	_ "github.com/donislawdev/TestingFilesGenerator/internal/format/pdf"
 	_ "github.com/donislawdev/TestingFilesGenerator/internal/format/png"
 	_ "github.com/donislawdev/TestingFilesGenerator/internal/format/txt"
+	_ "github.com/donislawdev/TestingFilesGenerator/internal/format/zip"
 )